requirements.txt
typer
sqlalchemy
sqlalchemy-utils
python-dotenv
psycopg2-binary
pandas
openpyxl